+++ CHANGES.md
@@ -11,6 +11,9 @@
 
 * Improvement: Basic layout of page
 * Support: Support MongoDB 4.0, 4.2 and 4.4
+* Support: Upgrade libs
+    * migrate-mongo
+    * mongoose
 
 ## v4.1.9